scripts/book-sync/requirements.txt
playwright>=1.40.0
python-frontmatter>=1.1.0
